fix: adicionar logica de semantica ao componente de empresas
